refactors dlthub cli module layout
