--- dlt/_workspace/cli/_workspace_command.py
@@ -1,161 +1,0 @@
-import argparse
-from typing import List, Optional
-
-from dlt.common.configuration.specs.pluggable_run_context import (
-    RunContextBase,
-)
-
-from dlt._workspace.cli import echo as fmt, utils
-from dlt._workspace._workspace_context import WorkspaceRunContext
-from dlt._workspace.typing import (
-    TCurrentProfileInfo,
-    TDeploymentManifestInfo,
-    TLocationInfo,
-    TProviderInfo,
-)
-from dlt._workspace.cli.utils import (
-    check_delete_local_data,
-    delete_local_data,
-    fetch_deployment_info,
-    fetch_workspace_info,
-)
-from dlt._workspace.cli._pipeline_command import list_pipelines
-
-
-def _format_location_tag(loc: TLocationInfo) -> str:
-    """Build a human-readable tag for a location (e.g. '(global, profile: dev)')."""
-    parts = []
-    if loc["scope"] == "global":
-        parts.append("global")
-    if profile_name := loc.get("profile_name"):
-        parts.append("profile: %s" % profile_name)
-    if not loc["present"]:
-        parts.append("not found")
-    if parts:
-        return " (%s)" % ", ".join(parts)
-    return ""
-
-
-def print_profile_section(
-    profile: TCurrentProfileInfo, configured_profiles: Optional[List[str]] = None
-) -> None:
-    """Renders the active-profile block: name, paths, pinned status, configured profiles."""
-    fmt.echo("Settings for profile %s:" % fmt.bold(profile["name"]))
-    fmt.echo("  Pipelines and other working data: %s" % fmt.bold(profile["data_dir"]))
-    fmt.echo("  Locally loaded data: %s" % fmt.bold(profile["local_dir"]))
-    if profile["is_pinned"]:
-        fmt.echo("  Profile is %s" % fmt.bold("pinned"))
-    if configured_profiles:
-        fmt.echo(
-            "Profiles with configs or pipelines: %s" % fmt.bold(", ".join(configured_profiles))
-        )
-
-
-def print_providers(providers: List[TProviderInfo], verbosity: int) -> None:
-    """Renders configuration provider locations; verbosity > 0 also lists not-found locations."""
-    fmt.echo(fmt.cli_cmd("found configuration in following locations:"))
-    total_not_found_count = 0
-    for prov in providers:
-        fmt.echo("* %s" % fmt.bold(prov["name"]))
-        for loc in prov["locations"]:
-            if loc["present"]:
-                tag = _format_location_tag(loc)
-                fmt.echo("    %s%s" % (loc["path"], tag))
-            else:
-                if verbosity > 0:
-                    tag = _format_location_tag(loc)
-                    fmt.echo("    %s" % fmt.style("%s%s" % (loc["path"], tag), fg="yellow"))
-                else:
-                    total_not_found_count += 1
-        if prov["is_empty"]:
-            fmt.echo("    provider is empty")
-    if verbosity == 0 and total_not_found_count > 0:
-        fmt.echo(
-            "%s location(s) were probed but not found. Use %s to see details."
-            % (fmt.bold(str(total_not_found_count)), fmt.bold("-v"))
-        )
-
-
-@utils.track_command("workspace", track_before=False, operation="info")
-def print_workspace_info(run_context: WorkspaceRunContext, verbosity: int = 0) -> None:
-    info = fetch_workspace_info()
-
-    if info["name"]:
-        fmt.echo("Workspace %s:" % fmt.bold(info["name"]))
-    fmt.echo("Workspace dir: %s" % fmt.bold(info["run_dir"]))
-    fmt.echo("Settings dir: %s" % fmt.bold(info["settings_dir"]))
-
-    profile = info["profile"]
-    if profile:
-        fmt.echo()
-        print_profile_section(profile, info["configured_profiles"])
-
-    fmt.echo()
-    print_providers(info["providers"], verbosity)
-    # installed toolkits
-    if info["installed_toolkits"]:
-        fmt.echo()
-        fmt.echo(
-            "Installed toolkits: %s"
-            % ", ".join(fmt.bold(n) for n in sorted(info["installed_toolkits"]))
-        )
-
-    # list pipelines in the workspace
-    fmt.echo()
-    list_pipelines(run_context.get_data_entity("pipelines"), verbosity)
-
-    # deployment manifest summary
-    fmt.echo()
-    _print_deployment_info(fetch_deployment_info(), verbosity)
-
-
-def _print_deployment_info(info: TDeploymentManifestInfo, verbosity: int) -> None:
-    status = info["status"]
-    if status == "not_found":
-        fmt.echo("Deployment: no manifest found (create __deployment__.py)")
-        return
-    if status == "generation_failed":
-        fmt.warning("Deployment: manifest generation failed")
-        if verbosity > 0 and info.get("error"):
-            fmt.echo(info["error"])
-        return
-
-    counts = info["counts_by_category"]
-    total = info["total_jobs"]
-    # sort by descending count, then by category name for stability
-    parts = [
-        "%d %s(s)" % (n, cat) for cat, n in sorted(counts.items(), key=lambda kv: (-kv[1], kv[0]))
-    ]
-    breakdown = ", ".join(parts)
-    fmt.echo("Deployment: %s job(s): %s" % (fmt.bold(str(total)), breakdown))
-
-    if verbosity == 0:
-        return
-
-    max_name = max((len(j["display_label"]) for j in info["jobs"]), default=8)
-    for job in info["jobs"]:
-        triggers: List[str] = []
-        if "default_trigger" in job:
-            triggers.append("🎯 %s" % job["default_trigger"])
-        triggers.extend(job["triggers"])
-        if not triggers:
-            if job["category"] in ("interactive", "mcp", "dashboard", "notebook"):
-                triggers = ["(interactive)"]
-            else:
-                triggers = ["(manual only)"]
-        fmt.echo("  %s  %s" % (job["display_label"].ljust(max_name), ", ".join(triggers)))
-
-
-@utils.track_command("workspace", track_before=False, operation="clean")
-def clean_workspace(run_context: RunContextBase, args: argparse.Namespace) -> None:
-    fmt.echo("Local pipelines data will be removed. Remote destinations are not affected.")
-    deleted_dirs = check_delete_local_data(run_context, args.skip_data_dir)
-    if deleted_dirs:
-        delete_local_data(run_context, deleted_dirs)
-
-
-@utils.track_command("dashboard", True)
-def show_workspace(run_context: WorkspaceRunContext, edit: bool) -> None:
-    from dlt._workspace.helpers.dashboard.runner import run_dashboard
-
-    run_dashboard(edit=edit)
