+++ dlt/_workspace/cli/commands.py
@@ -1,14 +1,12 @@
 import argparse
 import os
-import sys
-from typing import Callable, Dict, List, Optional, Tuple
+from typing import Callable, Dict, List, Optional
 
 from dlt._workspace.cli import echo as fmt, utils
 from dlt._workspace.cli import SupportsCliCommand, DEFAULT_VERIFIED_SOURCES_REPO
 from dlt.common.configuration.plugins import TCliCommandCompose
-from dlt._workspace.deployment.typing import DEFAULT_DEPLOYMENT_MODULE
 from dlt._workspace.cli.exceptions import CliCommandException, CliCommandInnerException
-from dlt._workspace.cli.utils import add_mcp_arg_parser, make_mcp_run_flags
+from dlt._workspace.cli.utils import add_mcp_arg_parser
 from dlt._workspace.cli._urls import (
     DLT_INIT_DOCS_URL,
     DLT_PIPELINE_COMMAND_DOCS_URL,
@@ -718,286 +716,3 @@ the `dlt` Airflow wrapper (https://github.com/dlt-hub/dlt/blob/devel/dlt/helpers
                 branch=deploy_args.pop("branch"),
                 **deploy_args,
             )
-
-
-class AiCommand(SupportsCliCommand):
-    command = "ai"
-    help_string = "Use AI-powered development tools and utilities"
-    # docs_url =
-    description = (
-        "The `dlt ai` command provides commands to configure your LLM-enabled IDE and MCP server."
-    )
-
-    def configure_parser(self, ai_cmd: argparse.ArgumentParser) -> None:
-        self.parser = ai_cmd
-
-        ai_subparsers = ai_cmd.add_subparsers(
-            title="Available subcommands", dest="operation", required=False
-        )
-
-        # status command
-        ai_subparsers.add_parser(
-            "status",
-            help="Show AI setup status: dlt version, agent, toolkits, readiness checks",
-        )
-
-        # init command
-        from dlt._workspace.cli._urls import (
-            DEFAULT_AI_WORKBENCH_BRANCH,
-            DEFAULT_AI_WORKBENCH_REPO,
-        )
-
-        init_cmd = ai_subparsers.add_parser(
-            "init",
-            help="Install initial AI rules and skills for your AI coding agent",
-        )
-        init_cmd.add_argument(
-            "--agent",
-            choices=["claude", "cursor", "codex"],
-            default=None,
-            help="AI coding agent to install for. Auto-detected if omitted.",
-        )
-        init_cmd.add_argument(
-            "--location",
-            default=DEFAULT_AI_WORKBENCH_REPO,
-            help="Advanced. Git URL or local path to AI workbench repository.",
-        )
-        init_cmd.add_argument(
-            "--branch",
-            default=DEFAULT_AI_WORKBENCH_BRANCH,
-            help="Advanced. Git branch to fetch from.",
-        )
-        init_cmd.add_argument(
-            "--overwrite",
-            default=False,
-            action="store_true",
-            help="Overwrite existing files instead of skipping them.",
-        )
-
-        # secrets command group
-        secrets_cmd = ai_subparsers.add_parser(
-            "secrets",
-            help="Manage secrets files used by dlt",
-            description="List, view (redacted), or update secret files used by dlt providers.",
-        )
-        secrets_subparsers = secrets_cmd.add_subparsers(
-            title="Available subcommands", dest="secrets_operation", required=False
-        )
-        secrets_subparsers.add_parser(
-            "list",
-            help="List secret file locations from providers",
-        )
-        view_cmd = secrets_subparsers.add_parser(
-            "view-redacted",
-            help="Print secrets TOML with all values replaced by '***'",
-            description=(
-                "Without --path, shows the unified view merged from all project"
-                " secret files. With --path, shows that exact file."
-            ),
-        )
-        view_cmd.add_argument(
-            "--path",
-            default=None,
-            help="Show this exact file instead of the unified provider view",
-        )
-        update_cmd = secrets_subparsers.add_parser(
-            "update-fragment",
-            help="Merge a TOML fragment into the secrets file",
-        )
-        update_cmd.add_argument(
-            "fragment",
-            nargs="?",
-            default=None,
-            help="TOML fragment string to merge; reads from stdin if omitted",
-        )
-        update_cmd.add_argument(
-            "--path",
-            required=True,
-            help="Path to the secrets TOML file to write to",
-        )
-
-        # toolkit command group — verb-first form: `ai toolkit <verb> <name>`
-        toolkit_cmd = ai_subparsers.add_parser(
-            "toolkit",
-            help="Manage AI toolkit plugins (list, info, install)",
-        )
-        toolkit_sub = toolkit_cmd.add_subparsers(dest="toolkit_operation", required=False)
-
-        # shared parent with --location and --branch
-        toolkit_common = argparse.ArgumentParser(add_help=False)
-        toolkit_common.add_argument(
-            "--location",
-            default=DEFAULT_AI_WORKBENCH_REPO,
-            help="Advanced. Git URL or local path to toolkit repository.",
-        )
-        toolkit_common.add_argument(
-            "--branch",
-            default=DEFAULT_AI_WORKBENCH_BRANCH,
-            help="Advanced. Git branch to fetch toolkit from.",
-        )
-
-        toolkit_sub.add_parser(
-            "list",
-            help="List available toolkits",
-            parents=[toolkit_common],
-        )
-        info_cmd = toolkit_sub.add_parser(
-            "info",
-            help="Show toolkit contents and components",
-            parents=[toolkit_common],
-        )
-        info_cmd.add_argument("name", help="Toolkit name")
-        install_cmd = toolkit_sub.add_parser(
-            "install",
-            help="Install toolkit components into project",
-            parents=[toolkit_common],
-        )
-        install_cmd.add_argument("name", help="Toolkit name")
-        install_cmd.add_argument(
-            "--agent",
-            choices=["claude", "cursor", "codex"],
-            default=None,
-            help="AI coding agent to install for. Auto-detected if omitted.",
-        )
-        install_cmd.add_argument(
-            "--overwrite",
-            default=False,
-            action="store_true",
-            help="Overwrite existing files instead of skipping them.",
-        )
-        install_cmd.add_argument(
-            "--strict",
-            default=False,
-            action="store_true",
-            help="Fail on validation warnings (invalid frontmatter, etc.).",
-        )
-
-        # shared run flags — used by both `dlt ai mcp [flags]` and `dlt ai mcp run [flags]`
-        mcp_run_flags = make_mcp_run_flags(default_port=8000)
-
-        mcp_cmd = ai_subparsers.add_parser(
-            "mcp",
-            help="Run or install the dlt MCP server",
-            parents=[mcp_run_flags],
-        )
-        mcp_sub = mcp_cmd.add_subparsers(dest="mcp_operation", required=False)
-
-        mcp_sub.add_parser("run", help="Start the MCP server (default)", parents=[mcp_run_flags])
-
-        mcp_install_cmd = mcp_sub.add_parser(
-            "install",
-            help="Install MCP server config into the current project",
-        )
-        mcp_install_cmd.add_argument(
-            "--agent",
-            choices=["claude", "cursor", "codex"],
-            default=None,
-            help="AI coding agent to install for. Auto-detected if omitted.",
-        )
-        mcp_install_cmd.add_argument(
-            "--features",
-            nargs="*",
-            default=None,
-            help="MCP feature sets to include in the server config",
-        )
-        mcp_install_cmd.add_argument(
-            "--name",
-            default="dlt-workspace",
-            help="Server name in the MCP config (default: dlt-workspace)",
-        )
-        mcp_install_cmd.add_argument(
-            "--overwrite",
-            default=False,
-            action="store_true",
-            help="Overwrite existing server config instead of skipping.",
-        )
-
-    def execute(self, args: argparse.Namespace) -> None:
-        from dlt._workspace.cli._urls import (
-            DEFAULT_AI_WORKBENCH_BRANCH,
-            DEFAULT_AI_WORKBENCH_REPO,
-        )
-        from dlt._workspace.cli.dlthub.ai import (
-            ai_status_command,
-            ai_init_command,
-            ai_mcp_run_command,
-            ai_mcp_install_command,
-            ai_secrets_list_command,
-            ai_secrets_view_redacted_command,
-            ai_secrets_update_fragment_command,
-            ai_toolkit_install_command,
-            ai_toolkit_list_command,
-            ai_toolkit_info_command,
-        )
-
-        if args.operation == "status":
-            ai_status_command()
-        elif args.operation == "init":
-            ai_init_command(
-                agent=args.agent,
-                location=args.location,
-                branch=args.branch,
-                overwrite=args.overwrite,
-            )
-        elif args.operation == "secrets":
-            op = getattr(args, "secrets_operation", None)
-            if op == "view-redacted":
-                ai_secrets_view_redacted_command(path=args.path)
-            elif op == "update-fragment":
-                fragment = args.fragment or sys.stdin.read()
-                ai_secrets_update_fragment_command(fragment=fragment, path=args.path)
-            else:
-                ai_secrets_list_command()
-        elif args.operation == "toolkit":
-            tk_op = getattr(args, "toolkit_operation", None)
-            if tk_op == "list":
-                ai_toolkit_list_command(
-                    location=args.location,
-                    branch=args.branch,
-                )
-            elif tk_op == "info":
-                if not args.name:
-                    fmt.error("Toolkit name is required for 'info'.")
-                    raise CliCommandException()
-                ai_toolkit_info_command(
-                    name=args.name,
-                    location=args.location,
-                    branch=args.branch,
-                )
-            elif tk_op == "install":
-                if not args.name:
-                    fmt.error("Toolkit name is required for 'install'.")
-                    raise CliCommandException()
-                ai_toolkit_install_command(
-                    name=args.name,
-                    agent=args.agent,
-                    location=args.location,
-                    branch=args.branch,
-                    overwrite=args.overwrite,
-                    strict=args.strict,
-                )
-            else:
-                # default: list toolkits
-                ai_toolkit_list_command(
-                    location=getattr(args, "location", DEFAULT_AI_WORKBENCH_REPO),
-                    branch=getattr(args, "branch", DEFAULT_AI_WORKBENCH_BRANCH),
-                )
-        elif args.operation == "mcp":
-            mcp_op = getattr(args, "mcp_operation", None)
-            if mcp_op == "install":
-                ai_mcp_install_command(
-                    agent=args.agent,
-                    features=args.features,
-                    name=args.name,
-                    overwrite=args.overwrite,
-                )
-            else:
-                # default: run
-                ai_mcp_run_command(
-                    port=getattr(args, "port", 8000),
-                    stdio=getattr(args, "stdio", False),
-                    sse=getattr(args, "sse", False),
-                    features=getattr(args, "features", None),
-                )
-        else:
-            self.parser.print_usage()
