+++ dlt/_workspace/mcp/tools/toolkit_tools.py
@@ -11,8 +11,8 @@ from dlt._workspace.cli.dlthub.ai.utils import (
     fetch_workbench_toolkit_info,
     fetch_workbench_toolkits,
 )
+from dlt._workspace.cli.dlthub.ai.typing import TToolkitInfo, TWorkbenchToolkitInfo
 from dlt._workspace.mcp.context import with_mcp_tool_telemetry
-from dlt._workspace.typing import TToolkitInfo, TWorkbenchToolkitInfo
 
 
 @with_mcp_tool_telemetry()
