+++ dlt/_workspace/typing.py
@@ -1,31 +1,11 @@
-from typing import Any, Dict, List, Literal, NamedTuple, Optional
-from dlt.common.typing import NotRequired
+from typing import List, Literal, NamedTuple, Optional
 
 from dlt.common.configuration.providers.provider import ConfigProvider
 from dlt.common.storages.configuration import TSchemaFileFormat
-from dlt.common.typing import TypedDict
+from dlt.common.typing import NotRequired, TypedDict
 
 
 TLocationScope = Literal["project", "global"]
-
-TAiStatusWarning = Literal[
-    "not_initialized",
-    "no_init_toolkit",
-    "no_toolkits",
-    "mcp_unavailable",
-]
-
-
-class TAiStatusInfo(TypedDict):
-    """AI setup status: version, agent, toolkits, and readiness warnings."""
-
-    dlt_version: str
-    agent_name: Optional[str]
-    initialized: bool
-    has_init_toolkit: bool
-    toolkits: Dict[str, "TToolkitIndexEntry"]
-    warnings: List[TAiStatusWarning]
-    mcp_error: NotRequired[str]
 
 
 class TLocationInfo(TypedDict):
@@ -43,128 +23,6 @@ class TProviderInfo(TypedDict):
     name: str
     is_empty: bool
     locations: List[TLocationInfo]
-
-
-class TProfileInfo(TypedDict):
-    """A single profile with its status flags."""
-
-    name: str
-    description: str
-    is_current: bool
-    is_pinned: bool
-    is_configured: bool
-
-
-class TCurrentProfileInfo(TProfileInfo):
-    """The active profile, extending base with session directories."""
-
-    data_dir: str
-    local_dir: str
-
-
-class TCurrentProfileFullInfo(TCurrentProfileInfo):
-    """The active profile plus filtered provider locations for the verbose `profile info` view."""
-
-    providers: List[TProviderInfo]
-    configured_profiles: List[str]
-
-
-class TWorkbenchComponentInfo(TypedDict):
-    """A skill, command, or rule inside a workbench toolkit."""
-
-    name: str
-    description: str
-
-
-class TWorkbenchMcpServerInfo(TypedDict):
-    """MCP server definition inside a workbench toolkit."""
-
-    command: str
-    args: List[str]
-
-
-class TToolkitInfo(TypedDict):
-    """Core toolkit info from plugin.json + toolkit.json."""
-
-    name: str
-    version: str
-    description: str
-    tags: List[str]
-    dependencies: NotRequired[List[str]]
-    workflow_entry_skill: NotRequired[str]
-
-
-class TWorkbenchToolkitInfo(TToolkitInfo):
-    """Extends meta with structural data from directory scan."""
-
-    skills: List[TWorkbenchComponentInfo]
-    commands: List[TWorkbenchComponentInfo]
-    rules: List[TWorkbenchComponentInfo]
-    mcp_servers: NotRequired[Dict[str, TWorkbenchMcpServerInfo]]
-    has_ignore: bool
-
-
-class TToolkitIndexEntry(TToolkitInfo, total=False):
-    """Installed toolkit record. Inherits required meta fields, adds optional tracking."""
-
-    installed_at: str
-    agent: str
-    files: Dict[str, Any]
-    mcp_servers: List[str]
-
-
-class TWorkspaceInfo(TypedDict):
-    """Full workspace state returned by `fetch_workspace_info`."""
-
-    name: Optional[str]
-    run_dir: str
-    settings_dir: str
-    global_dir: str
-    profile: Optional[TCurrentProfileInfo]
-    configured_profiles: List[str]
-    providers: List[TProviderInfo]
-    dlt_version: str
-    dlthub_version: Optional[str]
-    initialized: bool
-    installed_toolkits: Dict[str, TToolkitIndexEntry]
-
-
-TDeploymentManifestStatus = Literal["ok", "not_found", "generation_failed"]
-
-
-class TDeploymentJobInfo(TypedDict):
-    """A single job entry in the deployment manifest summary."""
-
-    job_ref: str
-    display_label: str
-    category: str
-    default_trigger: NotRequired[str]
-    triggers: List[str]
-
-
-class TDeploymentManifestInfo(TypedDict):
-    """Summary of the workspace deployment manifest."""
-
-    status: TDeploymentManifestStatus
-    error: NotRequired[str]
-    total_jobs: NotRequired[int]
-    counts_by_category: NotRequired[Dict[str, int]]
-    jobs: NotRequired[List[TDeploymentJobInfo]]
-
-
-class TRunJobInfo(TypedDict):
-    """Resolved `workspace run` request — all data needed to launch the job."""
-
-    job_ref: str
-    display_label: str
-    trigger: str
-    trigger_humanized: str
-    launcher: str
-    run_id: str
-    entry_point: Dict[str, Any]
-    manifest_warnings: List[str]
-    refresh_warning: NotRequired[str]
-    profile_warning: NotRequired[str]
 
 
 class TSourceItem(TypedDict):
