+++ tests/workspace/cli/dlthub/ai/utils.py
@@ -16,7 +16,7 @@ from dlt._workspace.cli.dlthub.ai.utils import (
     DEFAULT_AI_WORKBENCH_REPO,
     compute_file_hash,
 )
-from dlt._workspace.typing import TToolkitIndexEntry, TToolkitInfo
+from dlt._workspace.cli.dlthub.ai.typing import TToolkitIndexEntry, TToolkitInfo
 
 # known toolkits in the repo (init is now visible)
 KNOWN_TOOLKITS = ["data-exploration", "init", "rest-api-pipeline", "dlthub-runtime"]
